requirements.txt
flask
gitpython
requests
bandit
google-generativeai
python-dotenv
google-adk
